+++ inlang/source-code/message-bundle-component/src/stories/inlang-selector-configurator.ts
@@ -1,12 +1,14 @@
 import { LitElement, css, html } from "lit"
-import { customElement, property } from "lit/decorators.js"
+import { customElement, property, state } from "lit/decorators.js"
 
 import SlDropdown from "@shoelace-style/shoelace/dist/components/dropdown/dropdown.component.js"
 import SlButton from "@shoelace-style/shoelace/dist/components/button/button.component.js"
 import SlSelect from "@shoelace-style/shoelace/dist/components/select/select.component.js"
 import SlOption from "@shoelace-style/shoelace/dist/components/option/option.component.js"
 import SlTag from "@shoelace-style/shoelace/dist/components/tag/tag.component.js"
-import type { Message } from "@inlang/sdk/v2"
+import { createMessage, type LanguageTag, type Message } from "@inlang/sdk/v2"
+import { addSelector } from "../helper/crud/selector/add.js"
+import upsertVariant from "../helper/crud/variant/upsert.js"
 
 // in case an app defines it's own set of shoelace components, prevent double registering
 if (!customElements.get("sl-dropdown")) customElements.define("sl-dropdown", SlDropdown)
@@ -58,6 +60,12 @@ export default class InlangSelectorConfigurator extends LitElement {
 				gap: 4px;
 				flex-wrap: wrap;
 			}
+			.actions {
+				width: 100%;
+				display: flex;
+				flex-direction: column;
+				gap: 4px;
+			}
 		`,
 	]
 
@@ -65,12 +73,132 @@ export default class InlangSelectorConfigurator extends LitElement {
 	inputs: string[] | undefined
 
 	@property()
-	message: Message | undefined
+	message?: Message | undefined
 
-	private _getPluralCategories = () => {
-		return this.message?.locale
-			? new Intl.PluralRules(this.message?.locale).resolvedOptions().pluralCategories
+	@property()
+	languageTag: LanguageTag | undefined
+
+	@property()
+	triggerMessageBundleRefresh: () => void = () => {}
+
+	@property()
+	addMessage: (newMessage: Message) => void = () => {}
+
+	@state()
+	private _input: string | undefined
+
+	private _getPluralCategories = (): string[] | undefined => {
+		return this.languageTag
+			? new Intl.PluralRules(this.languageTag).resolvedOptions().pluralCategories
 			: undefined
+	}
+
+	private _handleAddSelector = (addVariants: boolean) => {
+		if (this._input) {
+			if (!this.message && this.languageTag) {
+				// create selector in not present message
+				const newMessage = createMessage({ locale: this.languageTag, text: "" })
+
+				// add selector
+				addSelector({
+					message: newMessage,
+					selector: {
+						type: "expression",
+						arg: {
+							type: "variable",
+							name: this._input,
+						},
+						annotation: {
+							type: "function",
+							name: "plural",
+							options: [],
+						},
+					},
+				})
+
+				if (addVariants) {
+					this._addVariants({ message: newMessage, variantMatcherArrays: [] })
+				}
+
+				this.addMessage(newMessage)
+			} else if (this.message) {
+				// get variant matchers arrays
+				const _variants = structuredClone(this.message ? this.message.variants : [])
+				const _variantMatcherArrays = _variants.map((variant) => variant.match)
+
+				// add selector
+				addSelector({
+					message: this.message,
+					selector: {
+						type: "expression",
+						arg: {
+							type: "variable",
+							name: this._input,
+						},
+						annotation: {
+							type: "function",
+							name: "plural",
+							options: [],
+						},
+					},
+				})
+
+				// add plural options if present
+				// TODO: for now always present
+				if (addVariants) {
+					this._addVariants({ message: this.message, variantMatcherArrays: _variantMatcherArrays })
+				}
+			}
+
+			this.triggerMessageBundleRefresh()
+		}
+	}
+
+	private _addVariants = (props: { message: Message; variantMatcherArrays: string[][] }) => {
+		const _categories = this._getPluralCategories()
+
+		if (_categories) {
+			if (props.variantMatcherArrays && props.variantMatcherArrays.length > 0) {
+				for (const variantMatcherArray of props.variantMatcherArrays) {
+					for (const category of _categories) {
+						upsertVariant({
+							message: props.message,
+							variant: {
+								// combine the matches that are already present with the new category -> like a matrix
+								match: [...variantMatcherArray, category],
+								pattern: [
+									{
+										type: "text",
+										value: "",
+									},
+								],
+							},
+						})
+					}
+				}
+			} else {
+				for (const category of _categories) {
+					upsertVariant({
+						message: props.message,
+						variant: {
+							// combine the matches that are already present with the new category -> like a matrix
+							match: [category],
+							pattern: [
+								{
+									type: "text",
+									value: "",
+								},
+							],
+						},
+					})
+				}
+			}
+		}
+	}
+
+	override async firstUpdated() {
+		await this.updateComplete
+		this._input = this.inputs && this.inputs[0]
 	}
 
 	override render() {
@@ -80,7 +208,19 @@ export default class InlangSelectorConfigurator extends LitElement {
 					<slot></slot>
 				</div>
 				<div class="dropdown-container">
-					<sl-select label="Input" size="small" value=${this.inputs && this.inputs[0]}>
+					<sl-select
+						@sl-change=${(e: CustomEvent) => {
+							const inputElement = e.target as HTMLInputElement
+							this._input = inputElement.value
+						}}
+						@sl-show=${(e: CustomEvent) => {
+							const inputElement = e.target as HTMLInputElement
+							this._input = inputElement.value
+						}}
+						label="Input"
+						size="small"
+						value=${this.inputs && this.inputs[0]}
+					>
 						${this.inputs &&
 						this.inputs.map((inputs) => {
 							return html`<sl-option value=${inputs}>${inputs}</sl-option>`
@@ -97,7 +237,23 @@ export default class InlangSelectorConfigurator extends LitElement {
 							})}
 						</div>
 					</div>
-					<sl-button size="small" variant="primary">Add selector</sl-button>
+					<div class="actions">
+						<sl-button
+							@click=${() => {
+								this._handleAddSelector(false)
+							}}
+							size="small"
+							variant="primary"
+							>Add selector</sl-button
+						>
+						<sl-button
+							@click=${() => {
+								this._handleAddSelector(true)
+							}}
+							size="small"
+							>Add selector with default variants</sl-button
+						>
+					</div>
 				</div>
 			</sl-dropdown>
 		`