+++ inlang/source-code/paraglide/paraglide-js/src/services/file-handling/write-output.test.ts
@@ -86,8 +86,10 @@ const mockFs = (files: memfs.DirectoryJSON) => {
 			// @ts-ignore - memfs has the same interface as node:fs/promises
 			vi.spyOn(_memfs.promises, prop).mockImplementation(lixFs[prop])
 		} else {
-			// @ts-ignore - memfs has the same interface as node:fs/promises
-			vi.spyOn(_memfs.promises, prop)
+			if (prop in _memfs.promises) {
+				// @ts-ignore - memfs has the same interface as node:fs/promises
+				vi.spyOn(_memfs.promises, prop)
+			}
 		}
 	}
 	return _memfs.promises as NodeishFilesystem