+++ inlang/source-code/server/CHANGELOG.md
@@ -4,6 +4,46 @@
 
 ### Patch Changes
 
+- @inlang/badge@0.7.38
+- @inlang/plugin-message-format@2.2.0
+- @inlang/rpc@0.3.34
+- @inlang/telemetry@0.3.49
+- @inlang/website@null
+
+## null
+
+### Patch Changes
+
+- @inlang/website@null
+- @inlang/badge@0.7.37
+- @inlang/plugin-message-format@2.2.0
+- @inlang/rpc@0.3.33
+- @inlang/telemetry@0.3.48
+
+## null
+
+### Patch Changes
+
+- @inlang/badge@0.7.36
+- @inlang/plugin-message-format@2.2.0
+- @inlang/rpc@0.3.32
+- @inlang/telemetry@0.3.47
+- @inlang/website@null
+
+## null
+
+### Patch Changes
+
+- @inlang/badge@0.7.35
+- @inlang/plugin-message-format@2.2.0
+- @inlang/rpc@0.3.31
+- @inlang/telemetry@0.3.46
+- @inlang/website@null
+
+## null
+
+### Patch Changes
+
 - @inlang/badge@0.7.35
 - @inlang/plugin-message-format@2.2.0
 - @inlang/rpc@0.3.30